feat: allow registration of multiple users

+++ endpoints/categories/category.php
@@ -2,11 +2,10 @@
 require_once '../../includes/connect_endpoint.php';
 require_once '../../includes/inputvalidation.php';
 
-session_start();
-
 if (isset($_SESSION['loggedin']) && $_SESSION['loggedin'] === true) {
     if (isset($_GET['action']) && $_GET['action'] == "add") {
-        $stmt = $db->prepare('SELECT MAX("order") as maxOrder FROM categories');
+        $stmt = $db->prepare('SELECT MAX("order") as maxOrder FROM categories WHERE user_id = :userId');
+        $stmt->bindParam(':userId', $userId, SQLITE3_INTEGER);
         $result = $stmt->execute();
         $row = $result->fetchArray(SQLITE3_ASSOC);
         $maxOrder = $row['maxOrder'];
@@ -18,10 +17,11 @@ if (isset($_SESSION['loggedin']) && $_SESSION['loggedin'] === true) {
         $order = $maxOrder + 1;
 
         $categoryName = "Category";
-        $sqlInsert = 'INSERT INTO categories ("name", "order") VALUES (:name, :order)';
+        $sqlInsert = 'INSERT INTO categories ("name", "order", "user_id") VALUES (:name, :order, :userId)';
         $stmtInsert = $db->prepare($sqlInsert);
         $stmtInsert->bindParam(':name', $categoryName, SQLITE3_TEXT);
         $stmtInsert->bindParam(':order', $order, SQLITE3_INTEGER);
+        $stmtInsert->bindParam(':userId', $userId, SQLITE3_INTEGER);
         $resultInsert = $stmtInsert->execute();
     
         if ($resultInsert) {
@@ -42,10 +42,11 @@ if (isset($_SESSION['loggedin']) && $_SESSION['loggedin'] === true) {
         if (isset($_GET['categoryId']) && $_GET['categoryId'] != "" && isset($_GET['name']) && $_GET['name'] != "") {
             $categoryId = $_GET['categoryId'];
             $name = validate($_GET['name']);
-            $sql = "UPDATE categories SET name = :name WHERE id = :categoryId";
+            $sql = "UPDATE categories SET name = :name WHERE id = :categoryId AND user_id = :userId";
             $stmt = $db->prepare($sql);
             $stmt->bindParam(':name', $name, SQLITE3_TEXT);
             $stmt->bindParam(':categoryId', $categoryId, SQLITE3_INTEGER);
+            $stmt->bindParam(':userId', $userId, SQLITE3_INTEGER);
             $result = $stmt->execute();
 
             if ($result) {
@@ -71,9 +72,10 @@ if (isset($_SESSION['loggedin']) && $_SESSION['loggedin'] === true) {
     } else if (isset($_GET['action']) && $_GET['action'] == "delete") {
         if (isset($_GET['categoryId']) && $_GET['categoryId'] != "" && $_GET['categoryId'] != 1) {
             $categoryId = $_GET['categoryId'];
-            $checkCategory = "SELECT COUNT(*) FROM subscriptions WHERE category_id = :categoryId";
+            $checkCategory = "SELECT COUNT(*) FROM subscriptions WHERE category_id = :categoryId AND user_id = :userId";
             $checkStmt = $db->prepare($checkCategory);
             $checkStmt->bindParam(':categoryId', $categoryId, SQLITE3_INTEGER);
+            $checkStmt->bindParam(':userId', $userId, SQLITE3_INTEGER);
             $checkResult = $checkStmt->execute();
             $row = $checkResult->fetchArray();
             $count = $row[0];
@@ -85,9 +87,10 @@ if (isset($_SESSION['loggedin']) && $_SESSION['loggedin'] === true) {
                 ];
                 echo json_encode($response);
             } else {
-                $sql = "DELETE FROM categories WHERE id = :categoryId";
+                $sql = "DELETE FROM categories WHERE id = :categoryId AND user_id = :userId";
                 $stmt = $db->prepare($sql);
                 $stmt->bindParam(':categoryId', $categoryId, SQLITE3_INTEGER);
+                $stmt->bindParam(':userId', $userId, SQLITE3_INTEGER);
                 $result = $stmt->execute();
                 if ($result) {
                     $response = [
